feat(serverless-framework-lambda): Test for deployed binaries, upload generated archive and rename project
