+++ .github/workflows/test.yaml
@@ -400,7 +400,7 @@ jobs:
           - firebase-functions
           - azure-functions-linux
           - azure-functions-windows
-          - serverless-lambda
+          - serverless-framework-lambda
 
     env:
       START_TIME: ${{ needs.start-time.outputs.start-time }}
