+++ platforms-serverless/serverless-framework-lambda/index.ts
@@ -33,11 +33,23 @@ export async function handler() {
 
   const deleteManyUsers = await client.user.deleteMany({})
 
+  /*
+  // list all files deployed in Lambda to debug when tests are failing
+  const dirTree = require("directory-tree");
+  const tree = dirTree(process.env.LAMBDA_TASK_ROOT);
+  console.dir(tree, { depth: null });
+  */
+ 
+  // list all files in node_modules/.prisma/client
+  const fs = require('fs')
+  const files = fs.readdirSync(process.env.LAMBDA_TASK_ROOT + "/node_modules/.prisma/client")
+
   return {
     version: Prisma.prismaVersion.client,
     createUser,
     updateUser,
     users,
     deleteManyUsers,
+    files,
   }
 }
