+++ platforms-serverless/serverless-framework-lambda/package.json
@@ -3,13 +3,13 @@
   "license": "MIT",
   "devDependencies": {
     "aws-sdk": "2.887.0",
-    "prisma": "2.22.0-dev.17",
+    "prisma": "2.22.0-dev.19",
     "serverless": "2.35.0",
     "ts-node": "9.1.1",
     "typescript": "4.2.4"
   },
   "dependencies": {
-    "@prisma/client": "2.22.0-dev.17",
-    "@types/node": "13.13.48"
+    "@prisma/client": "2.22.0-dev.19",
+    "@types/node": "13.13.50"
   }
 }
