+++ platforms-serverless/serverless-framework-lambda/prepare.sh
