+++ platforms-serverless/serverless-framework-lambda/prisma/schema.prisma
@@ -1,5 +1,6 @@
 generator client {
   provider      = "prisma-client-js"
+  // second binaryTarget `rhel` is required for deploying to AWS Lambda
   binaryTargets = ["native", "rhel-openssl-1.0.x"]
 }
