+++ platforms-serverless/serverless-framework-lambda/run.sh
