+++ platforms-serverless/serverless-framework-lambda/serverless.yml
@@ -15,3 +15,9 @@ functions:
     name: e2e-tests-serverless-lambda
     environment:
       SERVERLESS_LAMBDA_PG_URL: ${env:SERVERLESS_LAMBDA_PG_URL}
+
+# only include the Prisma binary required on AWS Lambda while packaging
+package:
+  patterns:
+    - '!node_modules/.prisma/client/query-engine-*'
+    - 'node_modules/.prisma/client/query-engine-rhel-openssl-1.0.x'
