+++ platforms-serverless/serverless-framework-lambda/test.sh
