+++ platforms-serverless/serverless-framework-lambda/tsconfig.json
